Django_App/requirements.txt
asgiref==3.9.1
colorama==0.4.6
Django==5.2.6
djangorestframework==3.16.1
loguru==0.7.3
mysqlclient==2.2.7
sqlparse==0.5.3
tzdata==2025.2
win32_setctime==1.2.0
